+++ README.md
@@ -77,6 +77,7 @@ public class MainLauncher {
 	- [x] [Quartz](http://www.quartz-scheduler.org)
 - [ ] 模板引擎
 	- [ ] [Beetl](http://ibeetl.com/)
+	- [ ] [jetbrick-template](https://github.com/subchen/jetbrick-template-2x)
 	- [ ] Vecloity
 	- [ ] FreeMarker
 	- [ ] Thymeleaf
