+++ nutzboot-demo/nutzboot-demo-simple-dao/src/main/java/io/nutz/demo/simple/MainLauncher.java
@@ -1,16 +1,9 @@
 package io.nutz.demo.simple;
 
-import java.util.List;
-
 import org.nutz.boot.NbApp;
-import org.nutz.dao.Cnd;
 import org.nutz.dao.Dao;
-import org.nutz.dao.pager.Pager;
 import org.nutz.ioc.loader.annotation.Inject;
 import org.nutz.ioc.loader.annotation.IocBean;
-import org.nutz.mvc.annotation.At;
-import org.nutz.mvc.annotation.Ok;
-import org.nutz.mvc.annotation.Param;
 
 import io.nutz.demo.simple.bean.User;
 
@@ -19,18 +12,6 @@ public class MainLauncher {
     
     @Inject
     Dao dao;
-    
-    @Ok("raw")
-    @At("/user/count")
-    public long count() {
-        return dao.count(User.class);
-    }
-    
-    @Ok("json:full")
-    @At("/user/query")
-    public List<User> query(@Param("..")Pager pager) {
-        return dao.query(User.class, Cnd.orderBy().asc("age"), pager);
-    }
     
     public void init() {
         dao.create(User.class, true);
